+++ src/services/events/deleteEventWithScope.js
@@ -1,4 +1,4 @@
-const { deleteEvents } = require('../../queries/events/deleteEvents');
+const { deleteEvents, deleteEventsForScope } = require('../../queries/events/deleteEvents');
 
 const validateScopes = (scopes) => {
 	if (!scopes) {
@@ -11,6 +11,11 @@ const validateScopes = (scopes) => {
 	}
 };
 
+deleteAllEventsForScope = async (scope) => {
+	validateScopes(scope);
+	return deleteEventsForScope(scope);
+}
+
 deleteEventWithScope = async (eventId, scopes) => {
 	validateScopes(scopes);
 	return deleteEvents(eventId, Object.keys(scopes));
@@ -18,4 +23,5 @@ deleteEventWithScope = async (eventId, scopes) => {
 
 module.exports = {
 	deleteEventWithScope,
+	deleteAllEventsForScope
 };